% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_affine}
\alias{check_affine}
\title{Check whether affine is a 4x4 numerical matrix.}
\usage{
check_affine(affine)
}
\arguments{
\item{affine}{whatever, hopefully a 4x4 numerical matrix.}
}
\description{
Check whether affine is a 4x4 numerical matrix.
}
\note{
Stops if its not.
}
\keyword{internal}
