% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_coords}
\alias{check_coords}
\title{Check whether coords is a nx3 numerical matrix.}
\usage{
check_coords(coords)
}
\arguments{
\item{coords}{numeric matrix to check, expected to be nx3.}
}
\description{
Check whether coords is a nx3 numerical matrix.
}
\note{
Stops if its not.
}
\keyword{internal}
