% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_voxels}
\alias{check_voxels}
\title{Check whether voxels is a nx3 integer matrix.}
\usage{
check_voxels(voxels)
}
\arguments{
\item{voxels}{integer matrix to check, expected to be nx3.}
}
\description{
Check whether voxels is a nx3 integer matrix.
}
\note{
Stops if its not.
}
\keyword{internal}
