% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_fsaverage.R
\name{colin27_voxels_to_fsaverage}
\alias{colin27_voxels_to_fsaverage}
\title{Map Colin27 voxels of reference file to fsaverage coords and vertices.}
\usage{
colin27_voxels_to_fsaverage(
  voxels,
  surface = "white",
  fs_home = Sys.getenv("FREESURFER_HOME"),
  silent = TRUE
)
}
\arguments{
\item{voxels}{integer nx3 matrix, the IJK voxel indices in the 256x256x256 cortex mask reference file, 1-based.}

\item{surface}{character string, the fsaverage surface (brain mesh) to load. Must be a valid FreeSurfer surface name like 'white', 'pial', 'orig, 'inflated'.}

\item{fs_home}{character string, path to the FreeSurfer installation. Alternatively, a hemilist of \code{freesurferformats::fs.surface} instances like \code{surface = list("lh"=mysurflh, "rh"=mysurfrh)}. Used to find the surfaces, at \code{<fs_home>/subjects/fsaverage/surf/<hemi>.<surface>}, where hemi is 'lh' and 'rh'. Can be NULL if 'surface' is a hemilist of fs.surface instances.}

\item{silent}{logical, whether to suppress output messages in case of coords outside of cortex.}
}
\value{
see \code{mni152_coords_to_fsaverage}
}
\description{
The voxel indices are specific to the reference volume and thus this function is of limited use in general, but it serves as an example on how to achieve voxel mapping.
}
\note{
The voxel indices used by this function are specific to the cortex mask reference volume, so it is preferable to use the \code{} function. However, the code of this function illustrates how to get the fsaverage coords based on the IJK voxel indices for your own image, so take it as a demo.
}
\keyword{internal}
