% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaverage_to_coords.R
\name{coord_closest_vertex}
\alias{coord_closest_vertex}
\title{Find closest surface vertex to a point, using Euclidean distance.}
\usage{
coord_closest_vertex(coordinate, surfaces)
}
\arguments{
\item{coordinate}{\code{nx3} numerical matrix or vector of length 3, the query point coordinates.}

\item{surfaces}{hemilist of \code{fs.surface} instances}
}
\value{
a data.frame with columns named 'query_x', 'query_y', 'query_z', 'lh_closest_vertex', 'lh_distance', 'rh_closest_vertex', 'rh_distance', 'both_closest_vertex', 'both_distance', 'both_hemi'.
}
\description{
Find closest surface vertex to a point, using Euclidean distance.
}
\keyword{internal}
