% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{doapply.transform.mtx}
\alias{doapply.transform.mtx}
\title{Apply a spatial transformation matrix to the given coordinates.}
\usage{
doapply.transform.mtx(coords, mtx)
}
\arguments{
\item{coords}{nx3 (cartesian) or nx4 (homogeneous) numerical matrix, the input coordinates. If nx4, left as is for homogeneous notation, if nx3 (cartesian) a 1 will be appended as the 4th position.}

\item{mtx}{a 4x4 numerical transformation matrix}
}
\value{
the coords after applying the transformation. If coords was nx3, nx3 is returned, otherwise nx4.
}
\description{
Apply a spatial transformation matrix to the given coordinates.
}
\note{
This function is copied from the develop version of freesurferformats. It will be called from there once the function is in the release.
}
