% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaverage_to_coords.R
\name{mni305_coords_to_mni152_coords}
\alias{mni305_coords_to_mni152_coords}
\title{Find MNI152 coordinate of fsaverage vertex closest to the given MNI305 coordinate.}
\usage{
mni305_coords_to_mni152_coords(
  coords,
  surface = "orig",
  fs_home = Sys.getenv("FS_HOME"),
  simplify = FALSE
)
}
\arguments{
\item{coords}{nx3 numerical matrix, the MNI305 query coordinates.}

\item{surface}{a character string defining the fsaverage surface to load (like \code{"white"} or \code{"orig"}), or a pre-loaded hemilist of surfaces (i.e., \code{freesurferformats::fs.surface} instances)}

\item{fs_home}{character string, path of the FreeSurfer directory from which the fsaverage surfaces should be loaded. Ignored if \code{surface} is a hemilist (in that case the surfaces have already been loaded).}

\item{simplify}{logical, whether to return a vector instead of a single-row matrix in case only a single query coordinate is given.}
}
\value{
nx3 numerical matrix, the MNI152 coordinates for the vertices closest to the given MNI305 query coordinates. Depending on the distance to the closest vertex, this may be way off. Also see the 'simplify' parameter.
}
\description{
Find MNI152 coordinate of fsaverage vertex closest to the given MNI305 coordinate.
}
\examples{
\dontrun{
  # Get MNI305 coords of fsaverage vertex 9092 (right hemisphere, orig surface).
  fsaverage_home = file.path(Sys.getenv("FS_HOME"), "subjects");
  rh_orig = freesurferformats::read.fs.surface(
    file.path(fsaverage_home, "fsaverage", "surf", "rh.orig"));
  mni305_pt = rh_orig$vertices[9092L, ];
  mni152_coords = mni305_coords_to_mni152_coords(mni305_pt);
}

}
