% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_to_fsaverage.R
\name{vol_to_fsaverage}
\alias{vol_to_fsaverage}
\title{Project or map values from MNI volume to fsaverage surface.}
\usage{
vol_to_fsaverage(
  input_img,
  template_type,
  rf_type = "RF_ANTs",
  interp = "linear",
  out_type = "curv",
  out_dir = "."
)
}
\arguments{
\item{input_img}{3D or 4D NIFTI or MGZ image instance of type \code{fs.volume}, or a character string that will be interpreted as a file system path to such a volume that should be loaded with \code{freesurferformats::read.fs.volume}. If 4D, the 4th dimension is considered the time/subject dimension.}

\item{template_type}{character string, the source template or the space that your input image is in. One of 'MNI152_orig', 'Colin27_orig', 'MNI152_norm', 'Colin27_norm'.}

\item{rf_type}{the regfusion type to use, one of 'RF_ANTs' or 'RF_M3Z'.}

\item{interp}{interpolation method, currently only 'linear' is supported.}

\item{out_type}{character string, the format of the output files. One of the following: 'curv' for FreeSurfer curv format, 'mgz' for FreeSurfer MGZ format, 'gii' for GIFTI format.}

\item{out_dir}{character string, the path to a writable output directory. If \code{NULL}, the returned named list contains the projected data (instead of the path of the file it was written to) at the keys 'lh' and 'rh', and the parameter 'out_type' is ignored.}
}
\value{
named list of 2 character strings, the output files (for the 2 hemispheres) at keys 'lh' and 'rh'. See the documentation for parameter 'out_dir' if you want the data in R instead.
}
\description{
Applies the Wu et al. regfusion method to obtain surface coords, then interpolates values.
}
\examples{
\dontrun{
  # Project a 3D MNI152 volume to fsaverage surface space, get data in R.
  input_file = system.file("extdata", "testdata",
    "MNI_probMap_ants.central_sulc.nii.gz", package = "regfusionr");
  res = vol_to_fsaverage(input_file, template_type = 'MNI152_orig',
    rf_type = 'RF_ANTs', out_dir = NULL);
  # res$lh contains per-vertex data for the left hemisphere.
}

}
\author{
Tim Schäfer for the R version, Wu Jianxiao and CBIG for the original Matlab version.
}
