% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaverage_to_vol.R
\name{fsaverage_to_vol}
\alias{fsaverage_to_vol}
\title{Project or map per-vertex values from the fsaverage surface to the cortex voxels of an MNI volume.}
\usage{
fsaverage_to_vol(
  lh_input,
  rh_input,
  template_type = "MNI152_orig",
  rf_type = "RF_ANTs",
  interp = "linear",
  out_type = "mgz",
  out_dir = NULL,
  fsaverage_path = NULL
)
}
\arguments{
\item{lh_input}{numerical vector of per-vertex data for the left hemisphere of the template subject. Must 163842 values for the \code{fsaverage} template.}

\item{rh_input}{numerical vector of per-vertex data for the right hemisphere of the template subject. Must 163842 values for the \code{fsaverage} template.}

\item{template_type}{character string, the target template or the space that your output volume should be in. One of 'MNI152_orig', 'Colin27_orig', 'MNI152_norm', 'Colin27_norm'.}

\item{rf_type}{the regfusion type to use, one of 'RF_ANTs' or 'RF_M3Z'.}

\item{interp}{interpolation method, currently only 'linear' and 'nearest' are supported. The performance of the 'linear' method is currently quite bad, and it will be rewritten in \code{C++} when I find the time.}

\item{out_type}{character string, the format of the output files. One of the following: 'mgz' or 'mgh' for FreeSurfer MGZ/MGH format, 'nii' for NIFTI v1 format. Ignored unless out_dir is not NULL.}

\item{out_dir}{optional character string, the path to a writable output directory to which the output should be written as volume files. If \code{NULL}, no data is written to files (but the result is returned as described below). If out_dir is not \code{NULL}, the return value additionally contains the following keys: 'out_file' and the output file format at key 'out_format'.}

\item{fsaverage_path}{character string or NULL, the file system path to the fsaverage directory (NOT including the 'fsaverage' dir itself). If \code{NULL}, defaults to the return value of \code{fsbrain::fsaverage.path()} on the system. This path is used to read the spherical surface (both hemisphere meshes) of the template subject.}
}
\value{
named list with keys 'projected' and 'projected_seg', each of which holds a \code{256x256x256} array with the projected data. The data in 'projected_seg' is identical to the data in 'projected', with the exception that data values originating from the right hemisphere have been incremented by 1000. See \code{out_dir} parameter to easily write results to files. If out_dir is not \code{NULL}, the return value additionally contains the following keys: 'out_file' and the output file format at key 'out_format'.
}
\description{
Applies the Wu et al. regfusion method to obtain MNI volume coordinates, then interpolates values.
}
\note{
THIS FUNCTION IS CURRENTLY WORK-IN-PROGRESS AND NOT PART OF THE OFFICIAL API, DO NOT USE IT OR BE PREPARED FOR UN-ANNOUNCED BREAKING CHANGES ANY TIME.

This function requires the packages 'fsbrain' and 'haze', which are optional dependencies. The package 'fsbrain' can be installed from CRAN. For 'haze', see \code{https://github.com/dfsp-spirit/haze}.
}
\examples{
\dontrun{
  lh_input = rnorm(163842L, 3.0, 0.2);
  rh_input = rnorm(163842L, 3.0, 0.2);
  res = fsaverage_to_vol(lh_input, rh_input);
}

}
\author{
Tim Schäfer for the R version, Wu Jianxiao and CBIG for the original Matlab version.
}
