% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_data_file}
\alias{get_data_file}
\title{Get path to a data file required by the package code.}
\usage{
get_data_file(filename, subdir = NULL)
}
\arguments{
\item{filename}{character string, the filename}

\item{subdir}{character string, subdirectories (if any).}
}
\description{
Get path to a data file required by the package code.
}
\keyword{internal}
