% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_method.R
\name{linear_fsaverage_coords_to_MNI152_coords}
\alias{linear_fsaverage_coords_to_MNI152_coords}
\title{Transform MNI305 coords (FreeSurfer fsaverage surface) to MNI152 coordinates using the linear method.}
\usage{
linear_fsaverage_coords_to_MNI152_coords(vertex_coords)
}
\arguments{
\item{vertex_coords}{nx3 matrix of coordinates, e.g., typically from fsaverage surface vertices.}
}
\value{
nx3 numerical matrix if MNI152 coords.
}
\description{
This uses the 4x4 matrix from the FreeSurfer Coordinate Systems documentation.
}
\note{
This is the opposite of using the \cite{Wu et al.} approach: a linear transformation matrix is used. This approach is mainly implemented in this package to allow you to easily check the difference between the methods.
}
