% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_fsaverage.R
\name{mni152_coords_to_fsaverage}
\alias{mni152_coords_to_fsaverage}
\title{Map MNI152 coords to fsaverage coords.}
\usage{
mni152_coords_to_fsaverage(
  coords,
  surface = "white",
  fs_home = Sys.getenv("FS_HOME")
)
}
\arguments{
\item{coords}{nx3 numeric matrix, the source RAS coordinates in the input image which must be in MNI152 space.}

\item{surface}{character string, the fsaverage surface (brain mesh) to load. Must be a valid FreeSurfer surface name like 'white', 'pial', 'orig, 'inflated'.}

\item{fs_home}{character string, path to the FreeSurfer installation. Used to find the surfaces, at \code{<fs_home>/subjects/fsaverage/surf/<hemi>.<surface>}, where hemi is 'lh' and 'rh'.}
}
\value{
namd list with entries 'fsaverage_vertices': integer vector of fsaverage surface vertex indices, 'hemi': vector of hemi strings for the vertices, 'fsaverage_coords': nx3 numeric matrix of target coordinates.
}
\description{
Map MNI152 coords to fsaverage coords.
}
\note{
see standalone_scripts_for_MNI_fsaverage_coordinates_conversion/CBIG_RF_MNICoord2fsaverageVertex.m
}
