% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaverage_to_coords.R
\name{mni305_coords_to_mni152_coords}
\alias{mni305_coords_to_mni152_coords}
\title{Find MNI152 coordinate of fsaverage vertex closest to the given MNI305 coordinate.}
\usage{
mni305_coords_to_mni152_coords(
  coords,
  surface = "orig",
  fs_home = Sys.getenv("FS_HOME")
)
}
\arguments{
\item{coords}{nx3 numerical matrix, the MNI305 query coordinates.}

\item{surface}{a character string defining the fsaverage surface to load (like \code{"white"} or \code{"orig"}), or a pre-loaded hemilist of surfaces (i.e., \code{freesurferformats::fs.surface} instances)}

\item{fs_home}{character string, path of the FreeSurfer directory from which the fsaverage surfaces should be loaded. Ignored if \code{surface} is a hemilist (in that case the surfaces have already been loaded).}
}
\value{
the MNI152 coordinates for the vertices closest to the given MNI305 query coordinates. Depending on the distance to the closest vertex, this may be way off.
}
\description{
Find MNI152 coordinate of fsaverage vertex closest to the given MNI305 coordinate.
}
