% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_to_fsaverage.R
\name{project_data}
\alias{project_data}
\title{Perform coord transform and trilinear data interpolation.}
\usage{
project_data(data, affine, ras, interp = "linear")
}
\arguments{
\item{data}{3D or 3D source image values}

\item{affine}{4x4 double matrix, the ras2vox transformation matrix}

\item{ras}{nx3 double matrix, the source RAS coordinates}

\item{interp}{character string, the interpolation mode. Only 'linear' is currently supported. We should support 'nearest' as well.}
}
\value{
data values interpolated at the RAS coordinates.
}
\description{
Perform coord transform and trilinear data interpolation.
}
\keyword{internal}
