% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_fsaverage.R
\name{vol_coords_to_fsaverage}
\alias{vol_coords_to_fsaverage}
\title{Map MNI152 or Colin27 volume coords to fsaverage coords and vertices.}
\usage{
vol_coords_to_fsaverage(
  coords,
  surface = "white",
  fs_home = Sys.getenv("FS_HOME"),
  silent = TRUE,
  rf_type = "RF_ANTs",
  template_type = "FSL_MNI152"
)
}
\arguments{
\item{coords}{nx3 numeric matrix, the source RAS coordinates in the input image which must be in MNI152/Colin27 space. The coords must be within the cortex, otherwise the mapping makes no sense and \code{NaN} values are returned for the respective coords.}

\item{surface}{character string, the fsaverage surface (brain mesh) to load. Must be a valid FreeSurfer surface name like 'white', 'pial', 'orig, 'inflated'.}

\item{fs_home}{character string, path to the FreeSurfer installation. Alternatively, a hemilist of \code{freesurferformats::fs.surface} instances like \code{surface = list("lh"=mysurflh, "rh"=mysurfrh)}. Used to find the surfaces, at \code{<fs_home>/subjects/fsaverage/surf/<hemi>.<surface>}, where hemi is 'lh' and 'rh'. Can be NULL if 'surface' is a hemilist of fs.surface instances.}

\item{silent}{logical, whether to suppress output messages in case of coords outside of cortex.}

\item{rf_type}{the regfusion type to use, one of 'RF_ANTs' or 'RF_M3Z'.}

\item{template_type}{character string, the source template or the space that your input image is in. One of 'MNI152_orig', 'Colin27_orig', 'MNI152_norm', 'Colin27_norm'.}
}
\value{
named list with entries 'fsaverage_vertices': integer vector of fsaverage surface vertex indices, 'hemi': vector of hemi strings for the vertices, 'fsaverage_coords': nx3 numeric matrix of target coordinates, 'query_mni_coords': copy of input parameter coords, 'query_mni_voxels': the voxel indices at the query RAS coords.
}
\description{
Map MNI152 or Colin27 volume coords to fsaverage coords and vertices.
}
\examples{
\dontrun{
  mni_ras = c(60.0, 0.0, 10.0)
  res = vol_coords_to_fsaverage(mni_ras, surface = "white");
  res$fsaverage_vertices;   # 9092
}

}
\author{
Tim Schäfer for the R version, Wu Jianxiao and CBIG for the original Matlab version.
}
