% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_to_fsaverage.R
\name{vol_coords_to_fsaverage_coords}
\alias{vol_coords_to_fsaverage_coords}
\title{Map MNI coords to fsaverage coords.}
\usage{
vol_coords_to_fsaverage_coords(
  mni_coords,
  template_type = "MNI152_orig",
  rf_type = "RF_ANTs"
)
}
\description{
Map MNI coords to fsaverage coords.
}
