% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_to_fsaverage.R
\name{vol_to_fsaverage}
\alias{vol_to_fsaverage}
\title{Map values from MNI volume to fsaverage surface.

Applies the Wu et al. regfusion method to obtain surface coords, then interpolates values.}
\usage{
vol_to_fsaverage(
  input_img,
  out_dir = ".",
  template_type = "MNI152_orig",
  rf_type = "RF_ANTs",
  interp = "linear",
  out_type = "curv"
)
}
\arguments{
\item{input_img}{3D or 4D NIFTI or MGZ image. If 4D, the 4th dimension is considered the time/subject dimension.}

\item{out_dir}{character string, the path to a writeable output directory.}

\item{template_type}{character string, the source template}

\item{rf_type}{the regfusion type to use.}

\item{interp}{interpolation method}

\item{out_type}{character string, the format of the output files. One of the following: 'curv' for FreeSurfer curv format, 'mgz' for FreeSurfer MGZ format.}
}
\description{
Map values from MNI volume to fsaverage surface.

Applies the Wu et al. regfusion method to obtain surface coords, then interpolates values.
}
